# src/main/java/com/eneik/epidemiology/model/PageMetadata.java
package com.eneik.epidemiology.model;

import java.util.Objects;

public class PageMetadata {
    private int page;
    private int size;
    private long totalElements;
    private int totalPages;

    public PageMetadata() {
    }

    public PageMetadata(int page, int size, long totalElements, int totalPages) {
        this.page = page;
        this.size = size;
        this.totalElements = totalElements;
        this.totalPages = totalPages;
    }

    public int getPage() {
        return page;
    }

    public void setPage(int page) {
        this.page = page;
    }

    public int getSize() {
        return size;
    }

    public void setSize(int size) {
        this.size = size;
    }

    public long getTotalElements() {
        return totalElements;
    }

    public void setTotalElements(long totalElements) {
        this.totalElements = totalElements;
    }

    public int getTotalPages() {
        return totalPages;
    }

    public void setTotalPages(int totalPages) {
        this.totalPages = totalPages;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        PageMetadata metadata = (PageMetadata) o;
        return page == metadata.page &&
                size == metadata.size &&
                totalElements == metadata.totalElements &&
                totalPages == metadata.totalPages;
    }

    @Override
    public int hashCode() {
        return Objects.hash(page, size, totalElements, totalPages);
    }
}